ofz#64109 crash in CharClass::isLetterNumeric
there appears to be an extra portion created, because HookChar is
set, but there doesn't seem to be a reason for the HookChar so we
get an unwanted trailing portion.

It seems that the HookChar is set when processing the field, presumably
when we restore the SwTextFormatInfo text, idx and len we should also
restore its original HookChar state.

At least for this ruby text case there the field contents are drawn over
the body text, unlike others where the contents are drawn inline with
the body text, where we appear to need to keep the hook char.

Change-Id: Id48a3ef09cd10dcc91048f8c08207d2d105839d9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159642
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>