sw floattable, delete UI: fix undo/redo
Pressing the unfloat context menu item for a frame resulted in 2 undo
actions and trying to undo both even crashed.

One problem was that no SwNode was left in the TextFrame and then it was
deleted; but undo created a TextFrame with an empty SwNode in it.

Fix this by inserting an empty text node at the end of the TextFrame and
only then moving content from it.

The other trouble was that in case the new text node has no matching
frame during MoveNodeRange(), then the fly frame will be deleted twice:
once during MoveNodeRange() (because it has become empty) and once more
during DelLayoutFormat(). This can be avoided by moving AppendTextNode()
to its own layout action. Finally also group the (now) 3 undo actions
together, because it was a single action on the UI.

Change-Id: I95ce0abbab79a293ecb209579c105f96d284a1c9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159730
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins