Cleanup make files for Java/C++ SDK examples
This fixes some minor issues with the make files for Java/C++ SDK
examples in DevelopersGuide/FirstSteps.

For Java examples, an exatra "-f Example_Java.mk" was a leftover from a
previous change.

For C++ examples, 'make clean' was not working correctly, because of a
leftover from a previous change.

Both should be fixed now.

Change-Id: Id936824675a9bfd43674ca526bc97a998ed9daf8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159777
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>