sd: remote: disentangle class RemoteServer
RemoteServer was both an object and a thread instantiated for the IP
server, and a bunch of static methods and data used by all servers.

Split out a class IPRemoteServer.

Move Bluetooth setup to SdDLL::RegisterRemotes().

Also the BluetoothServer was accessing RemoteServer::sCommunicators but
never locked the corresponding RemoteServer::sDataMutex.

Change-Id: I3ff39e68e619af1e91697124a1352b5f20350a22
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159785
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>