officecfg,sd: add Office::Security::Net::AllowInsecureImpressRemoteWiFi
The IPRemoteServer listens on all interfaces and accepts connections
with a "PIN", which is transmitted in the clear (without encryption),
so it is not as secure as a naive user might assume and should probably
not be used in public/untrusted networks.

Make it possible to use the Impress remote with Bluetooth only, which
appears to be more secure, by adding a new setting which defaults to
disabled.

The DiscoveryService which does mDNS/ZeroConf/Bonjour/Renezvous/whatever
-it's-called-today should not be needed for Bluetooth.

Adapt the RemoteServer code to the new situation that there can be a
BluetoothServer but not a IPRemoteServer.

Also add a checkbox to the "Slide Show Settings" dialog.

Change-Id: I0cd3f23b616e23351f166bc9681b45786df8a26a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159786
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>