tdf#158307 android a11y: Auto-detect toolbar height
Instead of hard-coding a toolbar height for the
cell address and formula bar in Calc, use
`android:layout_height="wrap_content"`, so the height
is determined based on the font size.
This avoids the text getting cut off when a large
font size has been selected in Android's accessibility
settings.

Change-Id: I9bfbe2b7a1db39ac11d9ed0ab15290f974782489
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159809
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>