tdf#158126 UI: Scalable buttons for Impress navigation bar
On some panels, small icons of navigation bar are not
convenient to use. Currently there are 4 sizes available
for other icons: automatic, small, large and extra large.
It is useful to implement support for scalable buttons in
the Impress navigation bar. This setting can be located in
Slide show -> Slide Show Settings (Display section)

Change-Id: Ief8cc0cbf4eac98ee90bf3e993e4fc44cea0165e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159294
Tested-by: Jenkins
Reviewed-by: Samuel Mehrbrodt <samuel.mehrbrodt@allotropia.de>