sw floattable: add per-frame wrap-on-all-pages mode
Currently the wrap-on-all pages mode is off by default and a
document-level setting can enable it for all split flys.

Allowing this at a per-frame level (and not only per-doc level) is
suggested in <https://issues.oasis-open.org/browse/OFFICE-4150>. Given
that this floating table proposal is coming from us, it makes sense to
also support this additional attribute from the proposal, even if it's
not supported by Word.

Fix this by adding a new SwFormatWrapTextAtFlyStart property on fly
frames: this is meant to be enabled when wrap-on-all-pages is wanted for
this fly, but it's disabled at a per-doc level.

Unlike SwFormatFlySplit, this is meant to be easy at a layout level,
since SwFlyAtContentFrame::IsWrapOnAllPages() is already a per-frame
function, it is just backed with a per-doc setting.

Change-Id: I8f562102c2bc366a36c08895b681c4a2ab256bd9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159847
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins