sd: remote: release Transmitter mutex when sending
The Bluetooth communication appears to be astonishingly slow on my
system; the socket is deliberately cleared of O_NONBLOCK and what's even
worse is that the Mutex that is held while sending prevents the main
thread from adding new messages to the queues.

It looks like there should be no issue with releasing the Mutex while
sending, and if a separate Transmitter thread is used in the first place
it makes no sense to hold the Mutex while doing so.

Change-Id: Ic993e6c7a7799832c86fd4dd8c6ddad9dab1780b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159924
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>