Fix HtmlWriter::characters
This method may be called after an end(), and then it shouldn't try
to close a tag again. Also in this case the characters shouldn't be
considered in the next start()/characters()/end() run.

Clarify the meaning of the previous mbElementOpen, by renaming it
to mbOpeningTagOpen. The old mbCharactersWritten flag was redundant.

Change-Id: Ie0ac6ddd88be774853a1fc152742b51793af798b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159932
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>