android: Suggest file name for PDF export
When using the PDF export feature in Android Viewer,
suggest a file name for the PDF file that matches
the current display name (which is usually the
file name, or "untitled" for a newly created doc
that hasn't been saved yet).

This can be achieved by setting `Intent.EXTRA_TITLE`
for the `ACTION_CREATE_DOCUMENT`. [1]

The `DocumentsContract.EXTRA_INITIAL_URI` already
set previously already results in the same
directory as the doc being preselected in the
file chooser:

> Callers can set a document URI through
> DocumentsContract#EXTRA_INITIAL_URI to indicate the initial location of
> documents navigator. System will do its best to launch the navigator in
> the specified document if it's a folder, or the folder that contains the
> specified document if not.

Filling in the current file name was suggested
in a Google Play review comment for the app.

[1] https://developer.android.com/reference/android/content/Intent#ACTION_CREATE_DOCUMENT

Change-Id: Idbd4a89416089f927e0232ce65161b43059ca46d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159959
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
(cherry picked from commit 872d26d4a859ff72e1e13068b798854185960de0)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160031
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>