tdf#124480 android: Override icon color for dark theme
Add a `values-night/colors.xml` that can be used to
override the colors specified in `values/colors.xml`
specifically for the dark theme.

Specify a light color for `toolbar_foreground`,
which is the color used for (most) toolbar icons,
both for the top-level as well as the bottom
toolbar.

This makes most icons show up in proper contrast
when using dark mode. Others are not making use of
the color defined here and will have to be looked
into separately.

Change-Id: Ic699fd1cecc607bd5e748648113336fe045e72b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160070
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>