tdf#124480 android: Use themed color for more icons
These icons used in the toolbars were previously using a
hard-coded fill color.
Most of them were using black, but `ic_line` and `ic_rect`
used in the "Insert" tab of the bottom toolbar, were
previously using a lighter gray than the other items in there
for no apparent reason.

Switch all of them to use the themed `toolbar_foreground`
color that other icons are also using.

Together with

    Change-Id Ic699fd1cecc607bd5e748648113336fe045e72b1
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Nov 28 13:13:12 2023 +0100

        tdf#124480 android: Override icon color for dark theme

, this makes these icons use a light color when using
the dark system theme.

Given the icons are no longer hard-coded to be black,
the files with a "_black" suffix will be renamed in
a follow-up commit. (Separate commit to keep the
diff clearer.)

Change-Id: I2061878067d21dcd743ad6ca8d1018db323755d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160073
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>