tdf#124480 android: Use themed color for icon on start activity
Similar to how

        Change-Id I2061878067d21dcd743ad6ca8d1018db323755d6
        Author: Michael Weghorn <m.weghorn@posteo.de>
        Date:   Tue Nov 28 14:22:54 2023 +0100

            tdf#124480 android: Use themed color for more icons

did for the icons in the toolbars, use the themed color for
the directory icon in the start activity as well.
While at it, switch to using the same shade of gray that
the toolbar icons also use.
And rename the icon to drop the "_grey" suffix, since this
is now themed and no longer using a hard-coded color.

Change-Id: I9fcf51b712a01aab3b9754bcfbd817c89171a64f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160075
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>