tdf#158459: call DeleteAndJoin in strict reverse order
The crash was caused by removal of a paragraph break (pointed to by a
PaM returned by pDelPam->GetNext()), that resulted in merge of the node
pDelPam pointed to, with the previous node. The destructin of the node
made pDelPam dangle. In debug builds, an assertion failed in destructor
of SwContentIndexReg. I.e., since the code always called DeleteAndJoin
on pDelPam->GetNext(), the processing went from second-to-last deletion
towards the begin of the document, and the last deletion (represented
by pDelPam) was handled last, when GetNext call wrapped.

This change makes sure that the order is strict, from last to first.

Change-Id: I5cb7fe2f06d4138d3c445eeca8220f0a87a82626
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160158
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 9c22a72a2fc92146d24c6b673d9c07b1e6ff83f2)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160049
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>