LOK: fix setView language problem  
Fix some mis-localization problem by checking for wrong language  
at setView(...).

setView does not change current localization, if the view we want  
to set is the current view.

But in some cases the language - view is not in a consistent state.
Maybe the language changed but the view did not, or the current view  
changed without language change (I found examples for both).

Changed setView(...) so that it checks, if the current language
matches to the current view, and if it does not, then we set
the view, even if we want to set to the current view.        

This won't fix everything, but hopefully it helps a lot.
I think we should make sure that current view - current Language  
are always changed at the same time (or at least we make sure
they always match).

Change-Id: Ie177b9b55f7befcbcf7cd1f62e402700f0e1aa60
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160219
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Tested-by: Andras Timar <andras.timar@collabora.com>