Related: tdf#158450 use SfxTabPage::GetFrameWeld
to get the parent window.

SfxTabPage::GetDialogController

returns mpImpl->mpSfxDialogController

and that is set in SfxTabPage::SfxTabPage ctor with

mpImpl->mpSfxDialogController = dynamic_cast<SfxOkDialogController*>(m_pDialogController);

and GetDialogController returns that mpImpl->mpSfxDialogController

while nearly everything is derived from OKDialogController, not
everything, like this column dialog case is, so
mpImpl->mpSfxDialogController is null and it all fails.

This same page is inside format, page styles as "columns" but that's the
usual multi-page dialog derived from SfxOkDialogController, its just
this single page dialog that is different.

The original idea was to wait until this parent dialog is needed to
figure out what to use. We have this SfxTabPage::SetDialogController so
the dialog controller and page can theoretically change after the ctor,
though maybe that doesn't happen in this case.

Looking at SfxTabPage::GetFrameWeld that one always returns the current
parent (of m_pDialogController, not the dynamically casted
mpImpl->mpSfxDialogController) so in the best choice to use in general.

Change-Id: I4617212fb27fdeeb0afc41123f7bc1b73f495c32
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160223
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>