tdf#148540 Revert "tdf#109790 RTF import: keep remembering...
... paragraph style between \cell and \row"
This reverts 5.4.1 commit aaa6a5202a447fb4e86d5f016d8e79fbc34a3ed7,
and rtfexport7's tdf109790.rtf unit test still passes.
I also did a visual test, which looks good.

After \cell, we normally get more \cells, so an impending
\row is completely unexpected most of the time.
I'm not really sure why that patch was ever thought to be good.

The problem was that \pard was not removing the paragraph style
that was assigned to an earlier column. The end result seemed
innocent (no bad formatting noticed),
but that is probably based on other work
done in the meantime which allows the unit test to still pass
even after all of "its code" has been reverted.

[If this causes a regression, perhaps m_pLastCharacterContext
could be of value?]

Change-Id: Ide9b65f5e5fa39c21bac6d8ed354bb88e0bbefe5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160233
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>